Update projct structure
