--- Example/DApp/ActivePairingItem.swift
@@ -1,6 +1,0 @@
-//struct ActivePairingItem {
-//    let peerName: String
-//    let peerURL: String
-//    let iconURL: String
-//    let topic: String
-//}
