--- Example/DApp/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  DApp
-//
-//  Created by Admin on 19/01/2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
